moves encounterUtils to view dir
