+++ src/views/encounterUtils.ts
