+++ src/views/hiv-testing-service/tabs/positive-tracking.component.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { EncounterList } from "@ohri/openmrs-esm-ohri-commons-lib";
 import { POSITIVE_TRACKING_ENCOUNTER_TYPE } from "../../../constants";
-import { getData } from "../../../encounterUtils";
+import { getData } from "../../encounterUtils";
 import { moduleName } from "../../../index";
 
 const columns = [
