+++ src/views/index-case-testing/tabs/ict-general.component.tsx
@@ -4,7 +4,7 @@ import {
   ICT_GENERAL_ENCOUNTER_TYPE,
   MRN_NULL_WARNING,
 } from "../../../constants";
-import { getData } from "../../../encounterUtils";
+import { getData } from "../../encounterUtils";
 import { moduleName } from "../../../index";
 import styles from "../../../root.scss";
 import { fetchIdentifiers } from "../../../api/api";
