+++ src/views/pmtct/child/tabs/pmtct-child-final-outcome.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { EncounterList } from "@ohri/openmrs-esm-ohri-commons-lib";
 import { PMTCT_CHILD_FINAL_OUTCOME_ENCOUNTER_TYPE } from "../../../../constants";
-import { getData } from "../../../../encounterUtils";
+import { getData } from "../../../encounterUtils";
 import { moduleName } from "../../../../index";
 import { getPatientEncounters } from "../../../../api/api";
