+++ src/views/pmtct/child/tabs/pmtct-child-followup.component.tsx
@@ -5,7 +5,7 @@ import {
   EncounterListColumn,
 } from "@ohri/openmrs-esm-ohri-commons-lib";
 import { PMTCT_FOLLOWUP_ENCOUNTER_TYPE } from "../../../../constants";
-import { getData } from "../../../../encounterUtils";
+import { getData } from "../../../encounterUtils";
 import { moduleName } from "../../../../index";
 
 const PMTCTFollowupEncounterList: React.FC<{ patientUuid: string }> = ({
