+++ src/views/pmtct/mother/tabs/pmtct-mother-discharge.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { EncounterList } from "@ohri/openmrs-esm-ohri-commons-lib";
 import { PMTCT_MOTHER_DISCHARGE_ENCOUNTER_TYPE } from "../../../../constants";
-import { getData } from "../../../../encounterUtils";
+import { getData } from "../../../encounterUtils";
 import { moduleName } from "../../../../index";
 
 const columns = [
