+++ src/views/pmtct/mother/tabs/pmtct-mother-enrollment.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { EncounterList } from "@ohri/openmrs-esm-ohri-commons-lib";
 import { PMTCT_MOTHER_ENROLLMENT_ENCOUNTER_TYPE } from "../../../../constants";
-import { getData } from "../../../../encounterUtils";
+import { getData } from "../../../encounterUtils";
 import { moduleName } from "../../../../index";
 
 const columns = [
