+++ src/views/program-management/tabs/family-information.component.tsx
@@ -4,7 +4,7 @@ import {
   FAMILY_INFO_ENCOUNTER_TYPE,
   MRN_NULL_WARNING,
 } from "../../../constants";
-import { getData } from "../../../encounterUtils";
+import { getData } from "../../encounterUtils";
 import { moduleName } from "../../../index";
 import styles from "../../../root.scss";
 import { fetchIdentifiers } from "../../../api/api";
