+++ src/views/program-management/tabs/followup.component.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { EncounterList } from "@ohri/openmrs-esm-ohri-commons-lib";
 import { FOLLOWUP_ENCOUNTER_TYPE } from "../../../constants";
-import { getData } from "../../../encounterUtils";
+import { getData } from "../../encounterUtils";
 import { moduleName } from "../../../index";
 
 const columns = [
