+++ src/views/visits/visits-summary.component.tsx
@@ -4,7 +4,7 @@ import {
   EncounterListColumn,
 } from "@ohri/openmrs-esm-ohri-commons-lib";
 import { FOLLOWUP_ENCOUNTER_TYPE, MRN_NULL_WARNING } from "../../constants";
-import { getData } from "../../encounterUtils";
+import { getData } from "../encounterUtils";
 import { moduleName } from "../../index";
 import styles from "../../root.scss";
 import { fetchIdentifiers } from "../../api/api";
